requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.41
pydantic==1.8.2
python-decouple==3.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymysql==1.1.1
email-validator==2.1.1
python-dotenv==1.0.1
starlette==0.14.2
typing-extensions==4.13.2
asgiref==3.8.1 